Merging swift-DEVELOPMENT-SNAPSHOT-2019-05-20-a into tensorflow (#1615)
* [Breakpoint] Remove unused include

* [Initialization] Remove extra SwiftASTContext initialize and terminate

This is already being handled in SystemInitializerFull. It doesn't need to
happen in common, because other tools (like lldb-server) don't actually use
SwiftASTContext.

* [Initialization] Move SwiftREPL setup/teardown to SystemInitializerFull

REPL setup doesn't need to happen in SystemInitializerCommon, especially because
tools like lldb-server don't need the SwiftREPL.

* [Initialization] Remove unused link against lldbPluginExpressionParserSwift

* Update opaque_return_types test to have availability

* Update opaque_return_types test to have availability

* [Initialization] Remove unused includes from SystemInitializerCommon

* [Initialization] Remove extra ClangASTContext init

This is also being initialized in SystemInitializerFull (which is the right
place to do it).

* lldb-server: remove link against lldbInterpreter

This dependency is unused. Remove the extraneous link.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@359738 91177308-0d34-0410-b5e6-96231b3b80d8

* Initialization: remove ObjectContainer from Common

This restructures the initialization path to move the ObjectContainer
initialization into the *full* initialization path. This is not needed
for the lldb-server initialization path. This helps strip off ~1MiB
from the binary.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@359810 91177308-0d34-0410-b5e6-96231b3b80d8

* Initialization: correct macro usage

`_MSC_VER` indiciates that you are building with MSVC, not that you are building
for Windows.  Use `_WIN32` (which identifies Win32 and Win64).

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@359817 91177308-0d34-0410-b5e6-96231b3b80d8

* Initialization: move InstructionEmulation to full initialization

The debug server does not need to use the instruction emulation. This
helps reduce the size of the final lldb-server binary by another ~100K
(~1% savings).

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@360067 91177308-0d34-0410-b5e6-96231b3b80d8

* Make sure GetObjectDescription falls back to the Objective-C runtime.

This fixes an unintended regression introduced by
https://reviews.llvm.org/D61451 by making sure the Objective-C runtime
is also tried when the "correct" language runtime failed to return an
object description.

rdar://problem/50791055

Differential Revision: https://reviews.llvm.org/D62015

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@360929 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 109c11f1afcc6a82325aa1ab8d61ceef3e8fafcb)

* Factor out switch statement into a helper function (NFC)

This addresses post-commit review feedback for https://reviews.llvm.org/D62015.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@360930 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit e1817a8f1e60d86c0465e1eb713cd00a319880fc)

* Remove SwiftASTContext::isTrivialOptionSetType. (NFC)

To serve the SwiftOptionSet data formatter, SwiftASTContext had an
elaborate mechanism in place to determine whether the type is
"trivial" or not. Nowadays, for all practical intents and purposes
quering for the existence of the rawValue child, appears to have the
same effect and was done by the data formatter anyway.

rdar://problem/50735738

* [StackFrame] Prefer synthetic variable if asked.

If a type has a synthetic we should give it preference.
I believe this works fine for most of the types anwyay,
but integers are "special", as lldb treats them both
as primitive types or structures depending on the code
path in the data formatters code. Eventually we should
just have a consistent reprensentation for integers and
floats (i.e. represent them for what they really are, structs).
But until that happens, we need to have this hack in place.

Fixes:
(lldb) frame var patatino[0]
(_ArrayBuffer<Int>) patatino[0] = {
  _storage = (rawValue = 4 values)
}

to print:
(lldb) frame var patatino[0]
(Int) [0] = 1

<rdar://problem/50705707>

* [StackFrame] Add a comment to justify the workaround.

* Add a test for an OptionSet with a computed rawValue property.

* [EditLine] Check string pointers before dereferencing them.

<rdar://problem/50875178>

* [Test] Be less stingent about whitespace in SwiftREPL tests

Be more permissive when it comes to whitespace. Currently, if the regex
matches nothing, we still expect two spaces.

* [Target] Remove unused include from LanguageRuntime

* [Target] Remove unused method from LanguageRuntime

* [Target] Stop linking against lldbPluginObjCLanguage

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@360945 91177308-0d34-0410-b5e6-96231b3b80d8

* [crashlog] Use loads() instead of readPlistFromString() for python 3.

<rdar://problem/50903413>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@361087 91177308-0d34-0410-b5e6-96231b3b80d8